+++ prcommon/prcommon/static/comp/prcommon2/clippings/windowsettings.js
@@ -103,8 +103,11 @@ return declare("prcommon2.clippings.windowsettings",
 		this._windowid = windowid;
 		this._customerid = customerid;
 		
-		this._dashboardsettingsstandardsearchby.deleteItem(this._dashboardsettingsstandardsearchby._itemsByIdentity[2]);
-		this._dashboardsettingsstandardsearchby.deleteItem(this._dashboardsettingsstandardsearchby._itemsByIdentity[3]);
+		//this._dashboardsettingsstandardsearchby.deleteItem(this._dashboardsettingsstandardsearchby._itemsByIdentity[2]);
+		if (this._dashboardsettingsstandardsearchby._itemsByIdentity[3] && this._dashboardsettingsstandardsearchby._itemsByIdentity[3] != undefined)
+		{
+			this._dashboardsettingsstandardsearchby.deleteItem(this._dashboardsettingsstandardsearchby._itemsByIdentity[3]);
+		}
 
 		request.post('/clippings/dashboardsettings/get_for_edit',
 				utilities2.make_params({ data : {customerid:this._customerid, windowid:this._windowid}})).
@@ -368,7 +371,10 @@ return declare("prcommon2.clippings.windowsettings",
 	{
 		if (this.daterangeid.get("value") >= 6 )
 		{
-			this._groupby.deleteItem(this._groupby._itemsByIdentity[1]);
+			if (this._groupby._itemsByIdentity[1] && this._groupby._itemsByIdentity[1] != undefined)
+			{
+				this._groupby.deleteItem(this._groupby._itemsByIdentity[1]);
+			}
 			this._groupby.save();
 			if (this.groupbyid.get("value") == 1)
 			{
